Fixed file links in README, renamed evaluation to scoring for clarity.
